backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
alembic==1.12.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.24.1